Make console task first class
