+++ quality/src/main/kotlin/net/twisterrob/gradle/quality/QualityPlugin.kt
@@ -5,9 +5,9 @@ import net.twisterrob.gradle.common.AGPVersions
 import net.twisterrob.gradle.common.BaseExposedPlugin
 import net.twisterrob.gradle.common.registerTask
 import net.twisterrob.gradle.pmd.PmdPlugin
+import net.twisterrob.gradle.quality.tasks.ConsoleReportTask
 import net.twisterrob.gradle.quality.tasks.GlobalTestFinalizerTask
 import net.twisterrob.gradle.quality.tasks.HtmlReportTask
-import net.twisterrob.gradle.quality.tasks.ValidateViolationsTask
 import org.gradle.api.Project
 import org.gradle.kotlin.dsl.apply
 
@@ -24,7 +24,7 @@ class QualityPlugin : BaseExposedPlugin() {
 		project.apply<PmdPlugin>()
 
 		if (project.rootProject == project) {
-			project.tasks.register(REPORT_CONSOLE_TASK_NAME, ValidateViolationsTask::class.java)
+			project.tasks.register(REPORT_CONSOLE_TASK_NAME, ConsoleReportTask::class.java)
 			project.tasks.register(REPORT_HTML_TASK_NAME, HtmlReportTask::class.java)
 			if (AGPVersions.isAvailable) {
 				project.plugins.apply(LintPlugin::class.java)
